units 6 & 7 Notes

+++ lecture/unit_6/Movies_MongoDB/controllers/movie.controller.js
@@ -1,6 +1,9 @@
 const router = require("express").Router();
 // const movieModel = require("../models/movie.model");
 const Movie = require("../models/movie.model");
+// Way to validate session directly in the controller and endpoints:
+const validateSession = require("../middleware/validate-session");
+
 // module.exports = router;
 
 // Error Response Function
@@ -11,8 +14,9 @@ const errorResponse = (res, error) => {
 };
 // TODO POST
 // http://localhost:4000/movies/
+// ! Adding validate session to grab user._id from token to save to database
 
-router.post("/", async (req, res) => {
+router.post("/", validateSession, async (req, res) => {
   try {
     //1. Pull data from client (body)
     const { title, genre, rating, length, releaseYear } = req.body;
@@ -30,6 +34,7 @@ router.post("/", async (req, res) => {
       rating,
       length,
       releaseYear,
+      owner_id: req.user._id
     });
     //3. Use mongoose method to save to MongoDB
     const newMovie = await movie.save();
@@ -46,7 +51,8 @@ router.post("/", async (req, res) => {
   }
 });
 // TODO GET One
-router.get("/:id", async (req, res) => {
+//! Adding Validate Session
+router.get("/:id", validateSession, async (req, res) => {
   try {
     // grab parameters from url
     const { id } = req.params;
@@ -122,8 +128,8 @@ router.get("/:id", async (req, res) => {
         
         Hint: parameters within method are optional
 */
-
-router.get("/", async (req, res) => {
+//! Adding validate Session by passing it in as a param for the endpoint
+router.get("/", validateSession, async (req, res) => {
   try {
     // This endpoint will only return all movies, no req or params needed
     // Await all documents from the Movie collection
@@ -181,10 +187,17 @@ router.get("/genre/:genre", async (req, res) => {
   }
 });
 // TODO PATCH One
-router.patch("/:id", async (req, res) => {
+//! Adding validateSession to make sure the movie also belongs to the logged in user.
+router.patch("/:id", validateSession, async (req, res) => {
   try {
     //1. Pull value from parameter
-    const { id } = req.params;
+    // const { id } = req.params; // Commented out
+    // Create a filter to check both id from req.params & owner_id against id from the token
+    const filter = {
+      _id: req.params.id,
+      owner_id: req.user._id
+    };
+
 
     //2. Pull data from the body.
     const info = req.body;
@@ -196,7 +209,8 @@ router.patch("/:id", async (req, res) => {
 
     //* findOneAndUpdate(query, document, options)
     // returnOptions allows us to view the updated document
-    const updatedMovie = await Movie.findOneAndUpdate({_id: id}, info, returnOption);
+    const updatedMovie = await Movie.findOneAndUpdate(filter, info, returnOption);
+    // const updatedMovie = await Movie.findOneAndUpdate({_id: id}, info, returnOption);
     // const update = await Movie.findOneAndUpdate({_id: id,});
     // const update = await Movie.findOneAndUpdate({id: _id});
     // if (!update) throw new Error ("ID Not Found!");
@@ -211,13 +225,13 @@ router.patch("/:id", async (req, res) => {
 });
 // TODO Delete One
 //http://localhost:4000/movies/643c6251c06fc36a2ad85a93
-router.delete('/:id', async (req,res) => {
+router.delete('/:id', validateSession, async (req,res) => {
   try {
       //1. Capture ID
       const { id } = req.params;
 
       //2. Use delete method to locate and remove based off ID
-      const deleteMovie = await Movie.deleteOne({_id: id});
+      const deleteMovie = await Movie.deleteOne({_id: id, owner_id: req.user._id});
 
       //3. Respond to Client
       deleteMovie.deletedCount ?
